Refactor key generation methods
